+++ MyHebrewBible/MyHebrewBible.Client/Features/Parasha/TableOfContents/TocRecord.cs
@@ -1,5 +1,5 @@
 ﻿namespace MyHebrewBible.Client.Features.Parasha.TableOfContents;
 
-public record TocRecord(long SectionId, long GroupCount, long ScriptureID_Beg, string? VerseRange, long VerseCount);
+public record TocRecord(int SectionId, int GroupCount, int ScriptureID_Beg, string? VerseRange, int VerseCount);
 
 // Ignore Spelling: Toc
