+++ MyHebrewBible/MyHebrewBible.Client/Helpers/HebrewFormat.cs
@@ -15,7 +15,7 @@ public enum WordEnum
 public class HebrewFormat
 {
 	// ToDo: Compare dupe code to Components\Report\Enums\Helper.cs
-	public static string GetSpan(long wordEnumLong, string hebrew1, string hebrew2, string hebrew3)
+	public static string GetSpan(int wordEnumLong, string hebrew1, string hebrew2, string hebrew3)
 	{
 		WordEnum wordEnum = (WordEnum)wordEnumLong;
 
@@ -34,7 +34,7 @@ public class HebrewFormat
 
 	// If the last part of the word is a Maqqef, don't add a space to the end of the word
 
-	public static string CheckMaqqef(long wordEnumLong, string hebrew1, string hebrew2, string hebrew3)
+	public static string CheckMaqqef(int wordEnumLong, string hebrew1, string hebrew2, string hebrew3)
 	{
 		WordEnum wordEnum = (WordEnum)wordEnumLong;
