+++ MyHebrewBible/MyHebrewBible/Endpoints/GetWordPartByScriptureId.cs
@@ -4,7 +4,7 @@ namespace MyHebrewBible.Endpoints;
 
 public class WordPartByScriptureIdRequest
 {
-	public long ScriptureId { get; set; }
+	public int ScriptureId { get; set; }
 }
 
 public class GetWordPartByScriptureId : Endpoint<WordPartByScriptureIdRequest, IEnumerable<WordPartByScriptureId>>  //
@@ -44,15 +44,15 @@ public class GetWordPartByScriptureId : Endpoint<WordPartByScriptureIdRequest, I
 
 public class WordPartByScriptureId
 {
-	public long ScriptureID { get; set; }
-	public long WordCount { get; set; }
-	public long SegmentCount { get; set; }
-	public long WordEnum { get; set; }
+	public int ScriptureID { get; set; }
+	public int WordCount { get; set; }
+	public int SegmentCount { get; set; }
+	public int WordEnum { get; set; }
 	public string? Hebrew1 { get; set; }
 	public string? Hebrew2 { get; set; }
 	public string? Hebrew3 { get; set; }
 	public string? KjvWord { get; set; }
-	public long Strongs { get; set; }
+	public int Strongs { get; set; }
 	public string? Transliteration { get; set; }
 	public int? FinalEnum { get; set; }
 }
